backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langchain>=0.1.0
langchain-openai>=0.0.2
neo4j>=5.16.0
qdrant-client>=1.7.1
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
requests>=2.31.0
pydantic>=2.5.3
tiktoken>=0.5.2
langchain-google-genai>=0.0.5
